pulses : add tests for the manifest commands
