+++ tests/pulses/conftest.py
@@ -12,4 +12,19 @@
 from __future__ import (division, unicode_literals, print_function,
                         absolute_import)
 
+import enaml
+import pytest
+
+with enaml.imports():
+    from .contributions import PulsesContributions
+
 pytest_plugins = (str('ecpy_pulses.testing.fixtures'))
+
+
+@pytest.fixture
+def workbench(pulses_workbench):
+    """Simply register the contributions for testing.
+
+    """
+    pulses_workbench.register(PulsesContributions())
+    return pulses_workbench
