+++ tests/pulses/test_dependencies.py
@@ -15,7 +15,6 @@ from __future__ import (division, unicode_literals, print_function,
 from operator import getitem
 
 import pytest
-import enaml
 
 from ecpy.app.dependencies.api import BuildDependency
 from ecpy_pulses.pulses.sequences.base_sequences import BaseSequence
@@ -23,18 +22,6 @@ from ecpy_pulses.pulses.pulse import Pulse
 from ecpy_pulses.pulses.shapes.modulation import Modulation
 from ecpy_pulses.pulses.shapes.square_shape import SquareShape
 from ecpy_pulses.testing.context import TestContext
-
-with enaml.imports():
-    from .contributions import PulsesContributions
-
-
-@pytest.fixture
-def workbench(pulses_workbench):
-    """Simply register the contributions for testing.
-
-    """
-    pulses_workbench.register(PulsesContributions())
-    return pulses_workbench
 
 
 @pytest.fixture
